Added simlink from stat.log to stdout
Reason to do so is because the pinpoint collector logs some stuff into stat.log that isn't send to stdout.

But we want those logs and we do not want to change the application itself. So quickest and easiest way to do so is to link the file to stdout.